refactor: rename useBatchMorphoVaultDeposit to useBatchVaultDeposit, move to hooks/vaults
Provider-aware batch deposit hook serving both Morpho and Spark (the provider param drives the Spark referral dispatch via buildVaultDepositCall). Fixed the now-sibling ../vaults/types import to ./types. Legacy name kept as a re-export alias in the morpho barrel.
