+++ apps/webapp/src/hooks/index.ts
@@ -134,6 +134,7 @@ export type {
 
 // Provider-neutral vault core (Morpho + Spark)
 export {
+  useBatchVaultDeposit,
   useVaultRedeem,
   useVaultWithdraw,
   useErc4626VaultData,
