+++ apps/webapp/src/hooks/vaults/index.ts
@@ -27,6 +27,7 @@ export { useSparkVaultRate } from './spark/useSparkVaultRate';
 export { useSparkVaultResolvedRate } from './spark/useSparkVaultResolvedRate';
 export { useVaultRatesByAddress, type VaultRatesByAddressHook } from './useVaultRatesByAddress';
 export { computeVaultLimits, type VaultLimits, type VaultLimitsInput } from './computeVaultLimits';
+export { useBatchVaultDeposit } from './useBatchVaultDeposit';
 export { useVaultRedeem } from './useVaultRedeem';
 export { useVaultWithdraw } from './useVaultWithdraw';
 export {
